% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianFunc.R
\name{medianFunc}
\alias{medianFunc}
\title{Format for printing median (IQR) or (Range) by factor}
\usage{
medianFunc(data, nvar, gvar = NULL, digits = 1, IQR = TRUE)
}
\arguments{
\item{data}{Data frame where nvar and gvar can be found}

\item{nvar}{Character string of column in data of which medians will be
calculated}

\item{gvar}{Character string of column in data of which medians will be
grouped by}

\item{digits}{Number of decimals the numbers will be rounded to}

\item{IQR}{Logical: If TRUE then median (IQR) is printed.  If FALSE then
median (range) is printed}
}
\value{
Returns a string of the format median (IQR or range) for each of the
levels in gvar.  If gvar is null then median (IQR or range) is returned for
nvar.
}
\description{
Meant to be used in creating R matrix to be exported as a table
}
\details{
na.rm=TRUE is defaulted in the calculation of median and IQR or min and max
}
\examples{

outcome <- rnorm(20)
group <- factor(rep(c("A", "B"), each = 10))
d.frame <- data.frame(outcome, group)

medianFunc(d.frame, "outcome", "group", IQR = FALSE)
}
\seealso{
meanFunc
}
\author{
University of Wisconsin-Madison Biostatistics and Medical
Informatics Department, Scott Hetzel M.S.
}
