\name{r2lme}
\alias{r2lme}
\title{Calculates R-squared for lme model}
\description{Calculates R-squared for lme model}

\usage{
r2lme(obj)
}
\arguments{
  \item{obj}{lme object}
}

\value{
  R-square value
}

\details{
  implements R2 from Edwards, Muller, Wolfinger, Qaqish and
  Schabenberger, 2008.
  need the aod package for wald.test function
}

\seealso{}

\author{}

\examples{

}
