% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFunc.R
\name{tableFunc}
\alias{tableFunc}
\title{Format for printing frequency and percentages}
\usage{
tableFunc(data, fvar, gvar = NULL, level = NULL, digits = 1)
}
\arguments{
\item{data}{Data frame where nvar and gvar can be found}

\item{fvar}{Character string of column in data of factor data}

\item{gvar}{Character string of column in data of grouping data}

\item{level}{Character string of level of fvar to be reported.  If null then
all levels are printed}

\item{digits}{Number of decimals to be printed}
}
\value{
Returns a vector of strings of N (\%) from table(fvar,gvar) with
columns printed first.  If gvar is NULL then N (\%) of fvar is returned.
}
\description{
Meant to be used in creating R matrix to be exported as a table
}
\details{
table(fvar,gvar) and 100*prop.table(table(fvar,gvar),2) are used to create N
(\%) for each level of fvar by gvar.  If only 1 level is needed (i.e. Yes/No
variable only needs Yes to be reported) then level can specify which level
to print
}
\examples{

set.seed(784)
outcome <- factor(sample(c("No","Yes"),replace=TRUE,size=20))
group <- factor(rep(c("A","B"),each=10))
d.frame <- data.frame(outcome,group)

tableFunc(d.frame,"outcome","group",level="Yes")

}
\seealso{
chisqFunc
}
\author{
University of Wisconsin-Madison Biostatistics and Medical
Informatics Department, Scott Hetzel M.S.
}
