% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwAccrualBars.R
\name{uwAccrualBars}
\alias{uwAccrualBars}
\title{Bar plot of frequency of accural}
\usage{
uwAccrualBars(
  accData,
  dateName,
  trxName = NULL,
  interval = c("months", "days", "years"),
  startDate = NULL,
  currentDate = NULL,
  pTitle = "Accrual Over Time",
  yLab = NULL,
  xLab = NULL,
  yLim = NULL,
  pHoriz = FALSE,
  pBeside = TRUE,
  linedbars = FALSE,
  printBarVals = TRUE,
  Legend = FALSE,
  LegendLoc = "topright",
  LegendCex = 0.8,
  cex.names = 1,
  titleCex = 1,
  barNamesLas = 1,
  barNamesSeq = 1,
  barNamesAngle = 0
)
}
\arguments{
\item{accData}{Data frame which includes an accural date for each patient}

\item{dateName}{(string) Variable name in accData indicating the date of
accrual. Must be a variable with class type 'dates' from library(chron). The
function will try to coherece a non-dates variable into a m/d/y formatted
dates class}

\item{trxName}{(string) Optional string giving the name of the variable in
accData that indicates treatment assignment. If used, then accrual for each
time point is given separately for each treatment level. If NULL then any
possible treatment levels are collapsed and treated as one large group}

\item{interval}{(string) Indicator of which time interval is to be
presented.  Default is 'months'.  1 letter abbreviation is sufficient}

\item{startDate}{(date) Specify the start date of accrual.  Defaults to the
minimum date of dateName}

\item{currentDate}{(date) Specify the current date.  Defaults to the maximum
date of dateName}

\item{pTitle}{(string) Title of graphic}

\item{yLab}{(string) Y label of graphic}

\item{xLab}{(string) X label of graphic}

\item{yLim}{(numeric) Limits of y axis of graphic. Default is the default of
barplot}

\item{pHoriz}{(logic) horiz attribute for barplot().  TRUE prints bars
horizontally}

\item{pBeside}{(logic) beside attribute for barplot().  FALSE prints stacked
bars, TRUE prints side-by-side bars}

\item{linedbars}{(logic) TRUE for lined bars, FALSE for solid colored bars}

\item{printBarVals}{(logic) TRUE for the frequency of each bar to be plot on
top of the bars.  0 values are suppressed from being printed}

\item{Legend}{(logic) TRUE for Legend to be printed. Only feasible if
trxName is not NULL.}

\item{LegendLoc}{(string) Strings allotted in legend() to be used as a
location for the legend.  See help(legend) for further explanation.}

\item{LegendCex}{(numeric) Multiplicative factor of printing size of legend}

\item{cex.names}{(numeric) Multiplicative factor of printing size of legend
axis label}

\item{titleCex}{(numeric) Multiplicative factor of printing size of title}

\item{barNamesLas}{(0,1,2 or 3) Option for axis labels to be horizontal or
veritcal}

\item{barNamesSeq}{(integer) Number indicating which time labels to print.
This number is given as the 'by' attribute to seq().  So every barNameSeq
would be printed.  Default is 1, which will print all time labels.  Mostly
helpful when interval='d'}

\item{barNamesAngle}{(integer) Number between 0 and 360 that gives the angle
in which the axis names will be printed.}
}
\value{
invisible
}
\description{
Bar plot of frequency of accural by days, months, or years.  Mostly for DSMB
use to assess accural
}
\examples{


Dates <- c(
  "03/01/2008", "09/01/2007", "07/01/2007", "12/01/2006", "08/01/2007", "08/01/2007",
  "04/01/2007", "04/01/2007", "04/01/2007", "11/01/2006", "11/01/2006", "12/01/2006",
  "12/01/2006", "10/01/2006", "01/01/2007", "11/01/2006", "12/01/2006", "02/01/2007",
  "08/01/2007", "08/01/2006", "09/01/2006", "04/01/2007", "05/01/2007", "11/01/2006",
  "10/01/2006", "03/01/2007", "10/01/2007", "07/01/2007", "03/01/2007", "05/01/2006",
  "05/01/2006", "11/01/2006", "12/01/2006", "02/01/2007", "02/01/2007", "02/01/2007",
  "04/01/2007", "05/01/2007", "07/01/2007", "08/01/2007", "11/01/2007", "10/01/2007",
  "09/01/2007", "10/01/2007", "05/01/2007", "01/01/2008", "12/01/2006", "02/01/2007",
  "03/01/2007", "04/01/2007", "04/01/2007", "05/01/2007", "07/01/2007", "02/01/2007",
  "07/01/2007", "10/01/2007", "01/01/2008"
)

Date <- chron::chron(Dates, format = c(dates = "m/d/y"))
TRT <- rep(c("A", "B"), c(30, 27))

data <- data.frame(Date, TRT)

layout(matrix(c(1, 2, 3, 4), nrow = 2, byrow = TRUE))
uwAccrualBars(
  accData = data, dateName = "Date", trxName = NULL, pTitle = "Accural",
  yLab = "Months", xLab = "Number of Patients",
  interval = "m", yLim = NULL, pHoriz = TRUE, pBeside = TRUE,
  linedbars = FALSE, Legend = TRUE, LegendLoc = "right",
  LegendCex = 0.5, cex.names = 0.7, barNamesLas = 1
)

uwAccrualBars(
  accData = data, dateName = "Date", trxName = NULL, pTitle = "Test",
  yLab = "Number of Patients", xLab = "Day", startDate = "01/01/2005", currentDate = "12/31/2010",
  interval = "d", yLim = NULL, pHoriz = FALSE, pBeside = TRUE,
  linedbars = TRUE, Legend = FALSE, LegendLoc = "topleft",
  LegendCex = 0.6, cex.names = 0.8, barNamesSeq = 40
)

uwAccrualBars(
  accData = data, dateName = "Date", trxName = NULL, pTitle = "Bars",
  yLab = "Months", xLab = "Number of Subjects", printBarVals = FALSE,
  interval = "y", yLim = NULL, pHoriz = TRUE, pBeside = FALSE,
  linedbars = TRUE, Legend = FALSE, LegendLoc = "bottomleft",
  LegendCex = 0.7, cex.names = 0.9
)

uwAccrualBars(
  accData = data, dateName = "Date", trxName = NULL,
  interval = "m", yLim = NULL, pHoriz = FALSE, pBeside = FALSE,
  linedbars = FALSE, Legend = TRUE
)
}
\seealso{
uwBars
}
\author{
Scott J Hetzel MS.  UW-Madison Department of Biostatistics and
Medical Informatics. Contributions from Patrick Lenon. Frontier Science and
Technology Research Foundation
}
