% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwBox.R
\name{uwBox}
\alias{uwBox}
\title{Box Plot function embedded in other box plot functions}
\usage{
uwBox(
  data,
  trxName = NULL,
  metricName,
  yLab = NULL,
  xLab = NULL,
  yLim = NULL,
  boxWex = 0.75,
  plotMean = TRUE,
  pOutliers = TRUE
)
}
\arguments{
\item{data}{data frame with metricName and trxName as columns}

\item{trxName}{column name of treatment factor in data}

\item{metricName}{column name of numeric variable in data}

\item{yLab}{(boxplot)(string) Y-axis label}

\item{xLab}{(boxplot)(string) X-axis label}

\item{yLim}{(boxplot)(numeric vector) Y-axis range}

\item{boxWex}{(boxplot)(numeric) see boxplot}

\item{plotMean}{(logical) if TRUE a dot will be plotted for where the mean
is}

\item{pOutliers}{(logical) if TRUE outliers are plotted}
}
\value{
Returns a list of sample sizes (n), means (mean), standard
deviations (sd), minimums (min), first quartile (q25), medians (median),
third quartile (q75), maximums (max)
}
\description{
Function called by other box plot functions to actually create the box plot.
This allows for a consistent look throughout multiple different functions
}
\details{
This function is an adaptation of the traditional boxplot() so that box plot
functions in biostatrpts can be similar

uwBox is called by any biostatrpts function that creates box plots.
}
\examples{


TRT <- c(rep("A", 20), rep("B", 20), rep("C", 20))
metric <- c(rnorm(20, 2, 1), rnorm(20, 3, 1), rnorm(20, 4, 1))

dat <- data.frame(TRT, metric)

uwBox(dat, "TRT", "metric")
}
\author{
University of Wisconsin-Madison Biostatistics and Medical
Informatics Department, Scott Hetzel M.S.
}
