% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxFunc.R
\name{wilcoxFunc}
\alias{wilcoxFunc}
\title{Format for printing p-value from wilcoxon rank sum test}
\usage{
wilcoxFunc(data, nvar, gvar)
}
\arguments{
\item{data}{Data frame where nvar and gvar can be found}

\item{nvar}{Character string of column in data of numeric data}

\item{gvar}{Character string of column in data of grouping data}
}
\value{
Returns a string of the p-value from a wilcoxon rank sum test or
kruskal wallis test if gvar has more than 2 levels
}
\description{
Meant to be used in creating R matrix to be exported as a table
}
\details{
If the p-value is less than 0.0005 then '< 0.001' is returned otherwise
p-value rounded to 3 decimals is returned
}
\examples{

outcome <- rnorm(20)
group <- factor(rep(c('A','B'),each=10))
d.frame <- data.frame(outcome,group)

wilcoxFunc(d.frame,'outcome','group')

}
\seealso{
medianFunc
}
\author{
University of Wisconsin-Madison Biostatistics and Medical
Informatics Department, Scott Hetzel M.S.
}
